% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_table.R
\docType{data}
\name{catch_table}
\alias{catch_table}
\title{Fish Catch Data from Santa Barbara}
\format{A dataframe with 5 rows, 6 columns and 30 variables.
\itemize{
 \item Fish
 \item Ventura
 \item SB
 \item Mussel Shoals
 \item Montecito
 \item Carpinteria
 \item Goleta
 }
 @source Lauren Kaapcke}
\usage{
catch_table
}
\description{
Fish catch data for 6 fisheries in Santa Barbara for 1 week. Data includes the number of of fish caught in Ventura, Santa Barbara, Mussel Shoals, Montecito, Carpinteria and Goleta. Catches were collected for 5 types of fish: Tuna, Shark, Salmon, Steelhead and Cod.
}
\keyword{datasets}
