% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_growth.R
\name{fish_growth}
\alias{fish_growth}
\title{Prediction of fish growth rate by water temperature}
\usage{
fish_growth(T, a, b, c, d)
}
\arguments{
\item{T}{Temperature of the surrounding water in degrees celsius}

\item{a}{Zero order parameter}

\item{b}{First order parameter}

\item{c}{Second order parameter}

\item{d}{Third order parameter}
}
\value{
G Specific growth rate
}
\description{
This function predicts fish growth rate by temperature in Celsius.
}
\examples{
fish_growth(T = 0, a = -0.2425, b = 0.1519, c = 0.0552, d = -0.002931)
}
\references{
https://www.sciencedirect.com/science/article/abs/pii/S0044848607005182
}
\author{
Lauren Kaapcke
}
