% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_prices.R
\docType{data}
\name{fish_prices}
\alias{fish_prices}
\title{Market Price Data for Different Types of Fish}
\format{A data frame with 5 rows, 2 columns and 5 variables
 \itemize{
 \item Fish
 \item Price $/lb
 }
 @source Lauren Kaapcke}
\usage{
fish_prices
}
\description{
This data is a table of current market prices per pound for different types of fish.
}
\keyword{datasets}
