% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries_summary.R
\name{fisheries_summary}
\alias{fisheries_summary}
\title{Fisheries Summary}
\usage{
fisheries_summary(prices, catch_table, plot = F)
}
\arguments{
\item{prices}{A table with types of fish in rows and prices for each fish the column.}

\item{catch_table}{A table with the number of fish caught in each fishery location.}

\item{plot}{An indicator whether the user would like to see a graph of revenue by location as an output. Default is FALSE.}
}
\value{
A list of the top fish caught in each location, the revenue for each location, the total revenue of all locations and a graph of revenue by location.
}
\description{
This function summarizes the most caught fish and the revenue for each fishery location as well as the total revenue and a plot of revenue by location.
}
\author{
Lauren Kaapcke
}
