% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_temp.R
\name{max_min_temp}
\alias{max_min_temp}
\title{Max Min Temperature Function}
\usage{
max_min_temp(input)
}
\arguments{
\item{The}{input is an a list of temperatures.}
}
\value{
Returns the maximum and minimum temperatures of the input list
}
\description{
This function is intended to take a list of temperatures and return the max and min temperatures.
}
\author{
Robert Saldivar

Jennifer Truong
}
