% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_precip.R
\name{mean_precip}
\alias{mean_precip}
\title{Mean precipitation}
\usage{
mean_precip(input)
}
\arguments{
\item{a}{list of precipitation events}
}
\value{
the mean precipitation of the input precipitation events
}
\description{
This function is going to calculate the mean of precipitation
}
\author{
Robert Saldivar

Jennifer Truong
}
