% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgrow.R
\name{popgrow}
\alias{popgrow}
\title{Population Growth Rate with Harvest}
\usage{
popgrow(
  population,
  hunters,
  growth_rate = 0.05,
  carrying_capacity = 3000,
  attack_rate = 0.05,
  handling_time = 0.2
)
}
\arguments{
\item{population}{is the population size}

\item{hunters}{is the number of hunters harvesting the the population. To find th growth rate of the population with out harvest, set hunters to 0}

\item{growth}{rate is the rate the population grows, default = 0.05}

\item{carrying}{capacity is that the number of indiviuals of a population that a habitat can support, default = 2000}

\item{attack}{rate is the rate that hunters can attack and harvest an individual of the population, default = 0.05}

\item{handling}{time is the amount of time that it takes for a hunter to havest one individual before they they can harvest the next one, default = 0.2}
}
\value{
is the intrinsic growth rate of the population at a given size with a give harvesting rate.
}
\description{
This function determines the intrinsic population growth rate of a population, if that population is being harvested/hunted
.
}
\author{
Robert Saldivar

Nathan BUrola


Assignment #1 Details
Include at least 2 inputs and several parameters (CHECK)
Add at least two different kinds of error checking (CHECK)
Add some error checking (at least 2 different kinds) (CHECK)
Save your function in a .R file (CHECK)
Adding comments to the function (CHECK)
In RMarkdown document, generate sample data for function in two different ways (CHECK)
Apply data to function (CHECK)
Graph and add some text describing how to interpret results
}
