% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somedata.R
\docType{data}
\name{somedata}
\alias{somedata}
\title{somedata}
\format{this data is a list}
\source{
Robert Saldivar

\describe{this is a set of random numbers}
}
\usage{
somedata
}
\description{
This data is random numbers that Robert Saldivar has randomly typed
}
\keyword{datasets}
