% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_weather_data.R
\docType{data}
\name{noaa_weather_data}
\alias{noaa_weather_data}
\title{Noaa Weather Data}
\format{is a table and data frame}
\source{
https://www.ncei.noaa.gov/metadata/geoportal/
}
\usage{
noaa_weather_data
}
\description{
This is weather data from NOAA which include daily precipiation and daily temperatures for Santa Barbara
}
\keyword{datasets}
