% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_supply.R
\name{water_supply}
\alias{water_supply}
\title{Water Supply}
\usage{
water_supply(input)
}
\arguments{
\item{input}{water precipitation data from dataset}
}
\value{
Amount of water supply (in kg/m^3)
}
\description{
This function calculates water supply based on precipitation trends
}
\author{
Robert Saldivar

Jennifer Truong
}
